Create new Component SwipeContainer
