+++ src/App.js
@@ -1,14 +1,12 @@
 import React, { useEffect } from 'react'
-import styled from 'styled-components/macro'
-import SwipeableViews from 'react-swipeable-views'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 
 import Layout from './common/Layout'
 import SideList from './SideList/SideList'
-import Song from './Song/Song'
 import AddSong from './forms/AddSong'
 import ListButton from './navigation/ListButton'
 import ChangeKeyButton from './ChangeKey/ChangeKeyButton'
+import SwipeContainer from './SwipeContainer'
 
 import useSongs from './hooks/useSongs'
 import useSideLists from './hooks/useSideLists'
@@ -51,23 +49,22 @@ export default function App() {
     }
   }
 
-  let swipeableViewContent
-  if (sideListType === 'singleSetlist') {
-    const activeSetlistIndex = setlists.findIndex(
-      setlist => setlist._id === activeSetlist
-    )
-    handleSwipeableView(setlists[activeSetlistIndex].songs)
-  } else {
-    handleSwipeableView(songs)
-  }
-
   return (
     <Router>
       <Switch>
         <Route exact path="/">
           <Layout>
-            <SwipeContainer>{swipeableViewContent}</SwipeContainer>
-
+            <SwipeContainer
+              songs={songs}
+              sideListType={sideListType}
+              setlists={setlists}
+              activeSetlist={activeSetlist}
+              isLoading={isLoading}
+              swipeIndex={swipeIndex}
+              handleChangeIndex={handleChangeIndex}
+              keyCounter={keyCounter}
+              changeKeyDirection={changeKeyDirection}
+            />
             <ListButton
               toggleSideList={toggleSideList}
               isSideListShown={isSideListShown}
@@ -85,21 +82,21 @@ export default function App() {
               setSwipeIndex={setSwipeIndex}
               setSetlists={setSetlists}
               setKeyCounter={setKeyCounter}
-            ></SideList>
+            />
             <ChangeKeyButton
               direction="up"
               handleKeyChangeClick={() => handleKeyChangeClick('up')}
               keyCounter={keyCounter}
               setChangeKeyDirection={setChangeKeyDirection}
               isSideListShown={isSideListShown}
-            ></ChangeKeyButton>
+            />
             <ChangeKeyButton
               direction="down"
               handleKeyChangeClick={() => handleKeyChangeClick('down')}
               keyCounter={keyCounter}
               setChangeKeyDirection={setChangeKeyDirection}
               isSideListShown={isSideListShown}
-            ></ChangeKeyButton>
+            />
           </Layout>
         </Route>
       </Switch>
@@ -117,32 +114,6 @@ export default function App() {
 
   function handleChangeIndex(index) {
     setSwipeIndex(index)
-  }
-
-  function handleSwipeableView(songs) {
-    swipeableViewContent = isLoading ? (
-      'Loading ...'
-    ) : (
-      <SwipeableViews
-        index={swipeIndex}
-        onChangeIndex={handleChangeIndex}
-        enableMouseEvents
-        animateTransitions={true}
-      >
-        {songs.length > 0 ? (
-          songs.map(song => (
-            <Song
-              key={song._id}
-              song={song}
-              keyCounter={keyCounter}
-              changeKeyDirection={changeKeyDirection}
-            />
-          ))
-        ) : (
-          <EmptySong>no song</EmptySong>
-        )}
-      </SwipeableViews>
-    )
   }
 
   function handleKeyChangeClick(direction) {
@@ -192,14 +163,3 @@ export default function App() {
     }
   }
 }
-
-const SwipeContainer = styled.div`
-  align-self: flex-start;
-  overflow: visible;
-`
-
-const EmptySong = styled.div`
-  height: 100vh;
-  width: 100vh;
-  align-self: flex-start;
-`
